Fix issues with Lumos Indexer.

+++ get-testnet-checkpoint-data.sh
@@ -1,2 +1,11 @@
-curl -O https://s3.amazonaws.com/cdn.ckb.tools/snapshots/20210704-nervos-ckb-indexer-testnet-snapshot.7z -L
-7z x 20210704-nervos-ckb-indexer-testnet-snapshot.7z -ohttp://3.235.223.161:18116
+#!/usr/bin/env bash
+
+if [ ! -f ./20210704-nervos-ckb-indexer-testnet-snapshot.7z ]; then
+	curl -O https://s3.amazonaws.com/cdn.ckb.tools/snapshots/20210704-nervos-ckb-indexer-testnet-snapshot.7z -L
+fi
+
+if [ -f ./indexer-data ]; then
+	rm -r ./indexer-data
+fi
+
+7z x 20210704-nervos-ckb-indexer-testnet-snapshot.7z -oindexer-data -aoa
